implement first-pass rules and stepping
First pass at implementing the rules for game of life.
I'm storing cell data as an array of integers, 0 meaning dead, 1
meaning alive.

As you click the cycle button, the nextCellState s calculated via
calculateNextState which takes an array of 0's and 1's and returns a new
array of the next state.

Also implemented a getIndexForGridCoords function which takes grid
coordinates in the form (row, column) and returns the corresponding
index if it exists, or null if not.